+++ CHANGELOG.md
@@ -35,6 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - (SSPP-239): Local file storage is completely remove now and replaced by Cloud S3 storage
+- (SSPP-285): Implemented new rating model to track ratings on notes
+- (SSPP-285): Added note rating controller functions to the note controller that uses the `ratingsService` in conjunction with other note helpers
 
 ## [Released]